rm duplicate evals, move extract_github_commits & extract_collaborators to combination
