--- evals/extract/extractGithubStars.ts
@@ -1,55 +1,0 @@
-import { z } from "zod";
-import { EvalFunction } from "../types/evals";
-
-export const extract_github_stars: EvalFunction = async ({ modelName, logger }) => {
-  const { stagehand, initResponse } = await initStagehand({
-    modelName,
-    logger,
-  });
-
-  const { debugUrl, sessionUrl } = initResponse;
-
-  try {
-    await stagehand.page.goto("https://github.com/facebook/react");
-
-    const { stars } = await stagehand.extract({
-      instruction: "Extract the number of stars for the project",
-      schema: z.object({
-        stars: z.number().describe("the number of stars for the project"),
-      }),
-      modelName,
-    });
-
-    const expectedStarsString = await stagehand.page
-      .locator("#repo-stars-counter-star")
-      .first()
-      .innerHTML();
-
-    const expectedStars = expectedStarsString.toLowerCase().endsWith("k")
-      ? parseFloat(expectedStarsString.slice(0, -1)) * 1000
-      : parseFloat(expectedStarsString);
-
-    const tolerance = 1000;
-
-    const isWithinTolerance = Math.abs(stars - expectedStars) <= tolerance;
-
-    await stagehand.context.close().catch(() => {});
-    return {
-      _success: isWithinTolerance,
-      stars,
-      debugUrl,
-      sessionUrl,
-      logs: logger.getLogs(),
-    };
-  } catch (error) {
-    console.error("Error or timeout occurred:", error);
-    await stagehand.context.close().catch(() => {});
-    return {
-      _success: false,
-      error: JSON.parse(JSON.stringify(error, null, 2)),
-      debugUrl,
-      sessionUrl,
-      logs: logger.getLogs(),
-    };
-  }
-};
