+++ evals/extract/extract_snowshoeing_destinations.ts
